+++ src/components/ListResultTable/ListResultTable.jsx
@@ -3,6 +3,8 @@ import PropTypes from "prop-types";
 import {Component} from "react";
 import QueryResultList from "./QueryResultList/QueryResultList";
 
+import configManager from "../../configManager/configManager";
+
 /**
  * @param {object} props - the props passed down to the component
  * @returns {Component} custom List as defined by react-admin which either shows a loading indicator or the query results
@@ -18,8 +20,6 @@ function ListResultTable(props) {
     resource,
     sort,
     variables,
-    changeVariables,
-    submitted,
     ...rest
   } = props;
 
@@ -30,6 +30,8 @@ function ListResultTable(props) {
       }
     }
   });
+
+  const query = configManager.getQueryWorkingCopyById(resource);
 
   return (
     <ListBase
@@ -45,7 +47,7 @@ function ListResultTable(props) {
       sort={sort}
     >
       {isLoading && <Loading loadingSecondary={"The page is loading. Just a moment please."} />}
-      {!isLoading && <QueryResultList {...rest} changeVariables={changeVariables} submitted={submitted} />}
+      {!isLoading && <QueryResultList resource={resource} { ...rest } />}
     </ListBase>
   );
 }
@@ -59,9 +61,9 @@ ListResultTable.propTypes = {
   filterDefaultValues: PropTypes.object,
   perPage: PropTypes.number,
   queryOptions: PropTypes.object,
-  resource: PropTypes.string,
+  resource: PropTypes.string.isRequired,
   sort: PropTypes.object,
-  variables: PropTypes.object,
+  variables: PropTypes.object.isRequired
 };
 
 export default ListResultTable;