+++ src/components/ListResultTable/QueryResultList/QueryResultList.jsx
@@ -7,10 +7,10 @@ import TableHeader from "./TableHeader/TableHeader";
 import Button from '@mui/material/Button';
 import SearchOffIcon from '@mui/icons-material/SearchOff';
 import { SvgIcon, Box, Typography } from "@mui/material";
+import PropTypes from "prop-types";
 import CustomQueryEditButton from "../../Dashboard/CustomQueryEditor/customQueryEditButton";
 
 import configManager from "../../../configManager/configManager";
-const config = configManager.getConfig();
 
 /**
  * @param {object} props - the props passed down to the component
@@ -21,7 +21,7 @@ function QueryResultList(props) {
   
   const QueryTitle = resource.options.label;
   const { data } = useListContext(props);
-  const {changeVariables, submitted} = props;
+  const { resource, changeVariables, submitted} = props;
   const [values, setValues] = useState(undefined);
   useEffect(() => {
     if (data && data.length > 0) {
@@ -31,15 +31,18 @@ function QueryResultList(props) {
     }
   }, [data]);
 
+  const config = configManager.getConfig();
+  const query = configManager.getQueryWorkingCopyById(resource);
+
   return (
     <div style={{ paddingLeft: '20px' , paddingRight: '10px' }}>
       <Title title={config.title} />
       {resource.options.queryGroupId === 'cstm' && <CustomQueryEditButton queryID={resource.name}/>}
       {submitted && <Aside changeVariables={changeVariables}/> /*  Adding button to make a new query - top left corner */ } 
-      <Typography fontSize={"2rem"} mt={2} > {QueryTitle} </Typography>
+      <Typography fontSize={"2rem"} mt={2} > {queryTitle} </Typography>
       {values ?(
-          <ListView title=" " actions={<ActionBar />} {...props} >
-            <Datagrid header={<TableHeader config={config}/>} bulkActionButtons={false}>
+        <ListView title=" " actions={<ActionBar />} {...props} >
+            <Datagrid header={<TableHeader query={query}/>} bulkActionButtons={false}>
               {Object.keys(values).map((key) => {
                 return (
                   <GenericField
@@ -52,11 +55,17 @@ function QueryResultList(props) {
             </Datagrid>
           </ListView>
         ): 
-          <NoValuesDiplay/>
+          <NoValuesDisplay/>
       }
     </div>
   );
 }
+
+QueryResultList.propTypes = {
+  resource: PropTypes.string.isRequired,
+  changeVariables: PropTypes.func.isRequired,
+  submitted: PropTypes.bool.isRequired
+};
 
 /**
  *
@@ -84,7 +93,7 @@ const Aside = (props) => {
     </div>
 )}
 
-const NoValuesDiplay = () => {
+const NoValuesDisplay = () => {
   return(
     <div>
       <Box display="flex" alignItems="center" sx={{m:3}}>