+++ src/dataProvider/SparqlDataProvider.js
@@ -35,10 +35,12 @@ configManager.on('configChanged', onConfigChanged);
 
 export default {
   getList: async function getList(resource, { pagination, sort, filter, meta }) {
-  
-    const query = findQueryWithId(resource);
-    query.limit = pagination.perPage;
-    query.offset = (pagination.page - 1) * pagination.perPage;
+    // make a working copy of the query object found in the configuration, to prevent changing the configuration
+    // this copy is extended here
+    // rendering should occur based on this working copy
+    const query = configManager.getQueryWorkingCopyById(resource);
+    const limit = pagination.perPage;
+    const offset = (pagination.page - 1) * pagination.perPage;
     query.sort = sort;
 
     handleComunicaContextCreation(query);
@@ -53,6 +55,8 @@ export default {
     }
 
     let results = await executeQuery(query);
+    let totalItems = results.length;
+    results = results.slice(offset, offset + limit);
 
     if (Object.keys(filter).length > 0) {
       results = results.filter((result) => {
@@ -64,7 +68,7 @@ export default {
 
     return {
       data: results,
-      total: query.totalItems
+      total: totalItems
     };
   },
   getOne: async function getOne() {
@@ -97,44 +101,23 @@ export default {
 };
 
 /**
- * Finds a query with the given id in config.queries
- * @param {number} id - identifier of a query
- * @returns {object} the query element from the configuration
+ * Fetches the query file and builds the final query text.
+ * @param {object} query - the query object working copy
+ * @returns {string} the built query text
  */
-function findQueryWithId(id) {
-  return config.queries.find((query) => query.id === id);
-}
-
-/**
- * Fetches the query file from the given query and returns its text.
- * @param {object} query - the query element from the configuration
- * @returns {string} the text from the file location provided by the query relative to query location defined in the config file, modified for our needs.
- */
-async function fetchQuery(query) {
+async function buildQueryText(query) {
   try {
-    const parser = new Parser();
-   // const result = await fetch(`${config.queryFolder}${query.queryLocation}`);
-    let rawText = await configManager.getQueryText(query)
-    
-    
+    let rawText = await configManager.getQueryText(query);
+
     if (query.variableValues) {
       rawText = replaceVariables(rawText, query.variableValues);
     }
 
     query.rawText = rawText;
+    const parser = new Parser();
     const parsedQuery = parser.parse(rawText);
     if (!query.variableOntology) {
       query.variableOntology = findPredicates(parsedQuery);
-    }
-    if (parsedQuery.limit !== undefined && query.offset + query.limit > parsedQuery.limit) {
-      parsedQuery.limit = parsedQuery.limit - query.offset;
-    } else {
-      parsedQuery.limit = query.limit;
-    }
-    if (parsedQuery.offset) {
-      parsedQuery.offset += query.offset;
-    } else {
-      parsedQuery.offset = query.offset;
     }
     if (!parsedQuery.order && query.sort && query.sort.field !== "id") {
       const { field, order } = query.sort;
@@ -192,13 +175,13 @@ function findPredicates(query) {
 }
 
 /**
- * A function that executes a given query and processes every result.
- * @param {object} query - the query element from the configuration
+ * Executes the query in scope and processes every result.
+ * @param {object} query - the query object working copy
  * @returns {Array<Term>} the results of the query
  */
 async function executeQuery(query) {
   try {
-    query.queryText = await fetchQuery(query);
+    query.queryText = await buildQueryText(query);
     return handleQueryExecution(
       await myEngine.query(query.queryText, {
         ...generateContext(query.comunicaContext),
@@ -291,7 +274,6 @@ async function handleQueryExecution(execution, query) {
     const resultType = execution.resultType;
     if (execution.resultType !== "boolean") {
       const metadata = await execution.metadata();
-      query.totalItems = await countQueryResults(query);
       variables = metadata.variables.map((val) => {
         return val.value;
       });
@@ -300,51 +282,6 @@ async function handleQueryExecution(execution, query) {
   } catch (error) {
     throw new HttpError(error.message, 500);
   }
-}
-
-/**
- * Predict the total number of elements in the result of a query
- * @param {object} query - the query element from the configuration
- * @returns {number} the actual number of results in the query, if it were executed
- */
-async function countQueryResults(query) {
-  const parser = new Parser();
-  const parsedQuery = parser.parse(query.rawText);
-  const distinctInitial = parsedQuery.distinct;
-  const offsetInitial = parsedQuery.offset;
-  const limitInitial = parsedQuery.limit;
-  parsedQuery.queryType = "SELECT";
-  parsedQuery.distinct = false;
-  parsedQuery.offset = 0;
-  if (parsedQuery.limit) {
-    delete parsedQuery.limit;
-  }
-  parsedQuery.variables = [
-    {
-      expression: {
-        type: "aggregate",
-        aggregation: "count",
-        expression: { termType: "Wildcard", value: "*" },
-        distinct: distinctInitial
-      },
-      variable: { termType: "Variable", value: "totalItems" },
-    },
-  ];
-  const generator = new Generator();
-  const countQuery = generator.stringify(parsedQuery);
-  const bindings = await myEngine.queryBindings(countQuery, {
-    sources: query.comunicaContext.sources,
-    fetch: fetch,
-    httpProxyHandler: proxyHandler,
-  });
-  let totalItems = parseInt((await bindings.toArray())[0].get("totalItems").value);
-  if (offsetInitial) {
-    totalItems -= offsetInitial;
-  }
-  if (limitInitial && totalItems > limitInitial) {
-    totalItems = limitInitial;
-  }
-  return totalItems;
 }
 
 const queryTypeHandlers = {