+++ CHANGELOG.md
@@ -13,7 +13,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Removed the display "Loaded: x results" in the action bar to avoid confusion (#83).
+
 ### Fixed
+
+- Fixed the broken production version due to a missing package (#123).
+- Corrected the counting of the total number of items in the result list (#120).
+- Avoided extending query objects in the configuration (#126).
 
 ## [1.2.0] - 2024-05-07